Move Save/Load into the engine; unify World into WorldModule
- Game gains Save(basePath)/Load(basePath, space, onLoaded, providers...),
  covering ECS + arbitrary game state (via Resources.SaveState/
  LoadState, using S's BinaryMarshaler when present) in one call each,
  including the post-Load re-insert into gokg.Space. This used to live
  in the collision-demo example; now it's engine functionality.
  Game.OnSetup is removed — its only use was this re-insert, now
  internal to Game.Load.
- spatial.World + PopulateModule + StaticPopulateModule are unified
  into WorldModule, with Populate/PopulateStatic as fluent builder
  methods (queued, applied via SetupSystems) instead of eagerly-run
  free functions.
- examples/collision-demo updated to the above: EventHandler.save()
  owns resources.State().Saves++ before calling game.Save; startup
  branches on the save file's presence between game.Load and
  worldModule.Populate(...).
- Add a permanent regression test (position_saveload_test.go) guarding
  against the promoted-BinaryMarshaler data loss bug fixed upstream in
  goke/gokg (kinematics.Position's AccX/AccY were silently zeroed by
  every Save before that fix).

+++ physics/collisions/narrow_phase.go
@@ -172,9 +172,9 @@ func (n *NarrowPhase) solve(cb *goke.CmdBuf, solverIterations int) {
 
 			isStaticB := contact.B.Vel == nil
 			if mtv1, mtv2, ok := contact.calculateMtv(boxA, boxB, isStaticB); ok {
-				n.space.Translate(uint64(contact.A.Entity), &contact.A.Pos.AABB, mtv1)
+				n.space.Translate(contact.A.Entity, &contact.A.Pos.AABB, mtv1)
 				if !isStaticB {
-					n.space.Translate(uint64(contact.B.Entity), &contact.B.Pos.AABB, mtv2)
+					n.space.Translate(contact.B.Entity, &contact.B.Pos.AABB, mtv2)
 				}
 			}
 		}
